+++ CHANGES.md
@@ -4,7 +4,8 @@ This files lists the changes during the lifetime of this project.
 
 ## unreleased
 
-- ?: improve user and colleague models
+- 266: always create Colleague for Users
+- 266: move labels from User to Colleague
 - 262: make sidepanels the same between 2 pages
 - 262: (migration) change status from assignment to service: (CONCEPT, OPEN, GESLOTEN)
 - 251: fix that placement table shows org labels instead of org names